requirements.txt
datasets
faiss-cpu 
transformers 
torch 
langchain 
langchain-community
google.generativeai 
tqdm
tools
pymupdf